+++ lib/bali/rules.rb
@@ -2,6 +2,7 @@ require "forwardable"
 
 class Bali::Rules
   extend Bali::Statics::Authorizer
+  extend Bali::Statics::ScopeRuler
 
   class << self
     extend Forwardable
@@ -9,10 +10,16 @@ class Bali::Rules
     attr_writer :current_role
     attr_reader :ruler
 
+    def_delegators :inheritable_role, :scope, :scope
     def_delegators :inheritable_role, :can, :can
     def_delegators :inheritable_role, :cant, :cant
     def_delegators :inheritable_role, :cant_all, :cant_all
     def_delegators :inheritable_role, :can_all, :can_all
+  end
+
+  def self.for(record_class)
+    rule_maker_cls_str = "#{record_class}#{Bali.config.suffix}"
+    rule_maker_cls_str.safe_constantize
   end
 
   def self.model_class
@@ -24,14 +31,8 @@ class Bali::Rules
 
   def self.role(*role_names, &block)
     role_names.each do |role_name|
-      if Symbol === role_name || String === role_name || NilClass === role_name
-        role = ruler[role_name]
-
-        if role.nil?
-          role = Bali::Role.new(role_name)
-          ruler << role
-        end
-
+      if Bali::Role::IDENTIFIER_CLASSES.include?(role_name.class)
+        role = ruler.find_or_create_role role_name
         role.instance_eval(&block)
       else
         raise Bali::DslError, "Cannot define role using #{param.class}. " +