+++ spec/test_app/app/controllers/api/transactions_controller.rb
@@ -1,0 +1,9 @@
+module Api
+  class TransactionsController < ActionController::API
+    def index
+      user = User.find params[:user_id]
+      transactions = rule_scope(Transaction.all, user)
+      render plain: transactions.pluck(:id).sort.join(", ")
+    end
+  end
+end