+++ spec/test_app/routes.rb
@@ -11,9 +11,12 @@ TestApp::Application.routes.draw do
     end
   end
 
+  resources :transactions, only: [:index]
+
   if Rails::VERSION::MAJOR >= 5
     namespace :api do
       resources :users, only: [:show]
+      resources :transactions, only: [:index]
     end
   end
 end