+++ CHANGELOG.md
@@ -102,3 +102,4 @@
 1. Built-in `be_able_to` matcher for RSpec
 2. Renaming `Bali::Printer.pretty_print` to `Bali::Printer.printable`
 3. Uniform `can?` and `cant?` authorizization checker
+4. Adding `scope` block